27-4-2019

+++ app/Http/Controllers/Movie_Controller.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+use Illuminate\Support\Str;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -14,7 +15,7 @@ use App\Http\Requests\CreateMovieRequest;
 use App\Http\Requests\EditMovieRequest;
 use App\Category;
 use App\Countrie;
-
+use App\Contact;
 
 class Movie_Controller extends Controller
 {
@@ -25,6 +26,15 @@ class Movie_Controller extends Controller
      */
     public function index()
     { 
+        $messagecount =session('counter');
+        $indexmessage = Contact::orderBy('id', 'desc')->limit(3)->get();
+         $count = $indexmessage->count();
+        if($count > session('counter'))
+        {
+            session(['counter' => $count]);
+        }
+
+        
         if(Auth::user()->isadmin())
         {
             $movies=Movie::paginate(20);
@@ -36,7 +46,7 @@ class Movie_Controller extends Controller
             $movies=Movie::where('user_id',$id)->paginate(20);
             $role='layouts.WorkerMaster';
         }
-        return view('admin.movie.index',compact('movies','role'));
+        return view('admin.movie.index',compact('movies','role','indexmessage'));
 
     }
 
@@ -59,8 +69,20 @@ class Movie_Controller extends Controller
         $country = Countrie::pluck('name','id')->all();
         return view('admin.movie.create',compact('category','country','role'));
 
-         
-       
+    }
+
+    public function createbyurl()
+    {
+        
+        if(Auth::user()->isadmin())
+        {
+           $role='layouts.AdminMaster';
+        }
+        else
+        {
+            $role='layouts.WorkerMaster';
+        }
+        return view('admin.movie.addbyurl',compact('role'));
     }
 
     /**
@@ -127,6 +149,104 @@ class Movie_Controller extends Controller
     session()->flash('success', 'Movie was added successfully!');
 
           return redirect()->route('movie.create');
+    }
+
+    public function storebyurl(Request $request)
+    {
+        $countrylist = Countrie::all();
+
+        $input=$request->all();
+        $url="http://www.omdbapi.com/?i=".$input['imdb']."&apikey=1c2aadc3";
+        $ch = curl_init();
+        curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
+        curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+        curl_setopt($ch, CURLOPT_URL, $url);
+        $data = curl_exec($ch);
+        curl_close($ch);
+        $result = json_decode($data,true);
+       
+        $title = $result["Title"];
+        $year = $result["Year"];
+        $story = $result["Plot"];
+        $imageurl = $result["Poster"];
+        $rate = $result["imdbRating"];
+        $imdbid = $result["imdbID"];
+        $country = explode(",",$result["Country"]);
+        $genre = $result["Genre"];
+        $user=Auth::user()->id;
+        //add the image
+        @$rawImage = file_get_contents($imageurl);
+        if($rawImage)
+        {
+           $name=date('y_m_d_H_i_s').Str::random(10).".png";
+           $name2=date('y_m_d_H_i_s').Str::random(11).".png";
+           file_put_contents("img/".$name,$rawImage);
+           $picture = new Photo;
+           $picture['path'] = $name;
+           $picture['pathcr']=$name2;
+           $picture->save();
+           $imageid = $picture->id;
+        }
+        //add movie
+         $movie=new Movie;
+         $movie['user_id'] = $user;
+         $movie['photo_id'] = $imageid;
+         //loop for country
+        foreach($countrylist as $countrie)
+        {
+            if($countrie["name"] == $country[0] )
+            {
+                $movie['country_id'] = $countrie['id'];
+            }
+
+        }
+         //end loop for country
+         $movie['name'] = $title;
+         $movie['year'] = $year;
+         $movie['rate'] = $rate;
+         $movie['quality'] = $input['quality'];
+         $movie['story'] = $story;
+         $movie['trailer'] = $input['trailer'];
+         $movie->save();
+         //end add movie and grap movie id that just added
+         $movie_id=$movie->id;
+         // add category for that movie
+        
+                  
+      
+            
+        $category = Category::pluck('name','id')->all();
+
+        foreach ($category as $key => $value) {
+            
+            if (strpos($genre, $value) !== false) {
+
+                    $catego= new Category_movie; 
+                    $catego['movie_id']=$movie_id;
+                    $catego['category_id']=$key;
+                    $catego->save();
+            }
+         }
+
+       // watch server
+         $watchm=new Watchmovie;
+         $watchm['movie_id']=$movie_id;
+         $watchm['server1']=$imdbid;
+         $watchm->save();
+         
+         //download server
+         $downl=new Downloadmovie;
+         $downl['movie_id']=$movie_id;
+         $downl['server1']='null';
+         $downl['server2']='null';
+         $downl['server3']='null';
+         $downl->save();
+
+      //sesion added succefly
+       session()->flash('success', 'Movie was added successfully!');
+
+         return redirect()->route('movie.createbyurl');
+
     }
 
     /**
@@ -251,12 +371,28 @@ class Movie_Controller extends Controller
     public function destroy($id)
     {  
         $movie=Movie::findOrFail($id);
-        unlink(public_path().$movie->photo->file.$movie->photo->path);
-        unlink(public_path().$movie->photo->file.$movie->photo->pathcr);
-        Category_movie::where('movie_id',$id)->delete();
-        $movie->delete();
-        $watchm=Watchmovie::where('movie_id',$id)->delete();
-        $downl=Downloadmovie::where('movie_id',$id)->delete();
-        return redirect()->route('movie.index');
+        if(file_exists(public_path().$movie->photo->file.$movie->photo->path)){
+            if(file_exists(public_path().$movie->photo->file.$movie->photo->pathcr))
+            {
+                unlink(public_path().$movie->photo->file.$movie->photo->pathcr);
+            }      
+            unlink(public_path().$movie->photo->file.$movie->photo->path);    
+            Category_movie::where('movie_id',$id)->delete();
+            $movie->delete();
+            $watchm=Watchmovie::where('movie_id',$id)->delete();
+            $downl=Downloadmovie::where('movie_id',$id)->delete();
+            return redirect()->route('movie.index');
+        }
+        else{
+            Category_movie::where('movie_id',$id)->delete();
+            $movie->delete();
+            $watchm=Watchmovie::where('movie_id',$id)->delete();
+            $downl=Downloadmovie::where('movie_id',$id)->delete();
+            return redirect()->route('movie.index');
+
+        }
+
+
+       
     }
 }
